# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WAR ROOM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AdminUser {
  userId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: AdminUserRole;
  createdAt: string;
}

export type UserRoleInputRole = typeof UserRoleInputRole[keyof typeof UserRoleInputRole];


export const UserRoleInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserRoleInput {
  role: UserRoleInputRole;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AdminStatus {
  isAdmin: boolean;
}

export type SignalSide = typeof SignalSide[keyof typeof SignalSide];


export const SignalSide = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export interface Signal {
  id: string;
  symbol: string;
  side: SignalSide;
  entry: number;
  /** @nullable */
  stop?: number | null;
  /** @nullable */
  target?: number | null;
  confidence: number;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  pnlPct?: number | null;
  createdAt: string;
}

export type SignalInputSide = typeof SignalInputSide[keyof typeof SignalInputSide];


export const SignalInputSide = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export interface SignalInput {
  symbol: string;
  side: SignalInputSide;
  entry: number;
  stop?: number;
  target?: number;
  confidence: number;
}