# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WAR ROOM API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Admin operations
  - name: signals
    description: Signal log operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
  /admin/users/{userId}/role:
    patch:
      operationId: adminSetUserRole
      tags: [admin]
      summary: Set user role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
  /admin/users/{userId}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
  /admin/check:
    get:
      operationId: adminCheck
      tags: [admin]
      summary: Check if current user is admin
      responses:
        "200":
          description: Admin status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List recent signals
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a signal (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalInput"
      responses:
        "201":
          description: Signal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "403":
          description: Forbidden
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AdminUser:
      type: object
      required: [userId, email, role, createdAt]
      properties:
        userId:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
    UserRoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, user]
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    AdminStatus:
      type: object
      required: [isAdmin]
      properties:
        isAdmin:
          type: boolean
    Signal:
      type: object
      required: [id, symbol, side, entry, confidence, createdAt]
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [LONG, SHORT]
        entry:
          type: number
        stop:
          type: ["number", "null"]
        target:
          type: ["number", "null"]
        confidence:
          type: number
        outcome:
          type: ["string", "null"]
        pnlPct:
          type: ["number", "null"]
        createdAt:
          type: string
    SignalInput:
      type: object
      required: [symbol, side, entry, confidence]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [LONG, SHORT]
        entry:
          type: number
        stop:
          type: number
        target:
          type: number
        confidence:
          type: number