# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WAR ROOM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserRole } from './adminUserRole';

export interface AdminUser {
  userId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: AdminUserRole;
  createdAt: string;
}