# lib/api-zod/src/generated/types/signal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WAR ROOM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignalSide } from './signalSide';

export interface Signal {
  id: string;
  symbol: string;
  side: SignalSide;
  entry: number;
  /** @nullable */
  stop?: number | null;
  /** @nullable */
  target?: number | null;
  confidence: number;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  pnlPct?: number | null;
  createdAt: string;
}