# lib/api-zod/src/generated/types/signalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WAR ROOM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignalInputSide } from './signalInputSide';

export interface SignalInput {
  symbol: string;
  side: SignalInputSide;
  entry: number;
  stop?: number;
  target?: number;
  confidence: number;
}