# lib/api-zod/src/generated/types/userRoleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WAR ROOM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRoleInputRole } from './userRoleInputRole';

export interface UserRoleInput {
  role: UserRoleInputRole;
}